Avoid sampler updates being off by one draw
Currently there is a bug where sampler changes take effect on the
subsequent draw instead of the draw we want.

Steps to reproduce the bug:
- Go to the rendering code in the 03_drawing example
- Move the first draw of image2 (line 142) below
  `canvas.set_default_sampler()` (line 158)
- Run the example

After changing the code according to the instructions above:
The rotating white image (the first draw) that isn't moving should be
sharp because we set the sampler using nearest filtering. Because the
sampler updates are off by one draw, it is blurry.

The moving rotating white image (the second draw after changing the
code) should be blurry because its drawn after `set_default_sampler()`,
but it is sharp.